remove utils by consolidating into the single fasta_compute_length.py script to simplify Pulsar deployment
